New Blake2.cs: Inheriting HashAlgorithm and IDisposable
The old Blake2B CSharp implementation created a lot of objects.
If you have made a `new Blake2B`, you also made Blake2BHasher : Hasher,
which makes indirectly a HashAlgorithmAdapter : HashAlgorithm,
most noteably it makes a Blake2BCore object,
and it does make a Blake2BConfig and Blake2BTreeConfig object.

In the new Blake2B.cs implementation, you just make a `new Blake2B()` object, which
also directly inherits `System.Security.Cryptography.HashAlgorithm`,
and therefore also `System.IDisposable`.

You do now

	byte[] bytes = Encoding.UTF8.GetBytes(text);
	byte[] value;

	using (var hash = new Blake2B()) value = hash.ComputeHash(bytes);
