--- csharp/Blake2Sharp.CompressionCodeGen/Properties/AssemblyInfo.cs
@@ -1,47 +1,0 @@
-﻿
-/// BLAKE2 reference source code package - C# implementation
-
-/// Written in 2012 by Christian Winnerlein  <codesinchaos@gmail.com>
-
-/// To the extent possible under law, the author(s) have dedicated all copyright
-/// and related and neighboring rights to this software to the public domain
-/// worldwide. This software is distributed without any warranty.
-
-/// You should have received a copy of the CC0 Public Domain Dedication along with
-/// this software. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("Blake2Sharp.CompressionCodeGen")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("Blake2Sharp.CompressionCodeGen")]
-[assembly: AssemblyCopyright("Copyright ©  2012")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("4af5636c-d52d-464f-a707-94464397988a")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
